refactor individual routes to separate files
