+++ .gitignore
@@ -1,3 +1,4 @@
 backend/node_modules
 backend/dist
 backend/.vscode
+.DS_Store
