+++ backend/src/routes/task/index.ts
@@ -1,0 +1,15 @@
+import { type FastifyInstance } from "fastify";
+
+import { addTaskCreateRoute } from "./create";
+import { addTasksGetRoute } from "./getMany";
+import { addTaskGetRoute } from "./getOne";
+import { addTaskDeleteRoute } from "./delete";
+import { addTaskUpdateRoute } from "./update";
+
+export function addTaskRoutes(fastify: FastifyInstance) {
+  addTaskCreateRoute(fastify);
+  addTasksGetRoute(fastify);
+  addTaskGetRoute(fastify);
+  addTaskDeleteRoute(fastify);
+  addTaskUpdateRoute(fastify);
+}
